Algeometry

Description

Classes

Linear space

class Eq b => LinSpace b el | el -> b where

The class representing a general linear space with basis of type b and element (vector) of type el.

Minimal complete definition

zero, isZero, monom, add, lmap, lapp, lfilter, assoc

Methods

zero :: el

The zero element.

isZero :: el -> Bool

Predicate for the zero element.

monom :: b -> Double -> el

Constructor for the monom.

isMonom :: el -> Bool

Predicate for the monom.

add :: el -> el -> el

Addition for two elements.

lmap :: (b -> Maybe (b, Double)) -> el -> el

Mapping of partial unary linear function through an element.

lapp :: (b -> b -> Maybe (b, Double)) -> el -> el -> el

Distribution of partial binary linear function through two elements.

lfilter :: (b -> Double -> Bool) -> el -> el

Extraction of elements with given predicate.

coeff :: b -> el -> Double

Scalar coefficient of an element

assoc :: el -> [(b, Double)]

Representation of an element as assoclist.

scale :: LinSpace b el => Double -> el -> el

Returns an element of a linear space, scaled by given factor.

elems :: LinSpace b el => el -> [b]

Returns list of basis elements for an element.

coefs :: LinSpace b el => el -> [Double]

Returns list of coefficients for an element.

terms :: LinSpace b el => el -> [el]

Returns list of monoms for an element.

lerp :: (Num el, LinSpace b el) => el -> el -> Double -> el

Linear interpolation for two elements, parameterized for interval [0,1].

Clifford algebras

class (Eq mv, Ord g, LinSpace [g] mv) => CliffAlgebra g mv where

The class representing Clifford algebra with generators of type g and element (multivector) of type mv.

Minimal complete definition

algebraSignature, square, generators

Methods

algebraSignature :: mv -> (Int, Int, Int)

The signature of Clifford algebra. The first argument is proxy needed to resolve the functional dependence for class instance.

square :: mv -> g -> Double

Squares of the generators, which define the Clifford algebra. The first argument is proxy needed to resolve the functional dependence for class instance.

generators :: [mv]

List of generators of the Clifford algebra.

grade :: mv -> Int

Returns the grade of a multivector.

geom :: mv -> mv -> mv

The geometric product.

outer :: mv -> mv -> mv

The outer product (same as (∧)).

rcontract :: mv -> mv -> mv

The right contraction (same as (|-)).

lcontract :: mv -> mv -> mv

The left contraction (same as (-|)).

inner :: mv -> mv -> mv

The inner product (same as (·)).

rev :: mv -> mv

The reverse of a multivector.

inv :: mv -> mv

The inversion of a multivector.

conj :: mv -> mv

The conjugate of a multivector

dual :: mv -> mv

The dual of a multivector.

rcompl :: mv -> mv

The right complement of a multivector.

lcompl :: mv -> mv

The left complement of a multivector.

(∧) :: CliffAlgebra g mv => mv -> mv -> mv infixr 9

The infix operator for the outer product

(|-) :: CliffAlgebra g mv => mv -> mv -> mv infix 8

The infix operator for rcontract.

(-|) :: CliffAlgebra g mv => mv -> mv -> mv infix 8

The infix operator for the lcontract.

(∙) :: CliffAlgebra g mv => mv -> mv -> mv infix 8

The infix operator for the inner product

(•) :: CliffAlgebra g mv => mv -> mv -> Double infix 9

The infix operator for the scalar product

Elements

e :: CliffAlgebra g mv => g -> mv

Returns a generator of the algebra.

e_ :: CliffAlgebra g mv => [g] -> mv

Returns a monomial element of the algebra (outer product of generators).

scalar :: CliffAlgebra g mv => Double -> mv

Returns a scalar element of the algebra.

>>> scalar (sin 1) + scalar (cos 1) * e12 :: VGA 2
0.8414709848078965 + 0.5403023058681398e₁₂

kvec :: GeomAlgebra g mv => Int -> [Double] -> mv

Returns a k-vector with given grade and coefficients.

>>> kvec 1 [1,2,3] :: VGA 2
e₁ + 2.0e₂ + 3.0e₃
>>> kvec 2 [1,2,3] :: VGA 3
e₁₂ + 2.0e₁₃ + 3.0e₂₃

avec :: GeomAlgebra g mv => Int -> [Double] -> mv

Returns a normalized k-antivector with given grade and coefficients.

>>> avec 1 [1,2,3] :: VGA 3
3.0e₁₂ + 2.0e₁₃ + e₂₃
>>> avec 2 [1,2,3] :: VGA 3
3.0e₁ + 2.0e₂ + e₃

nvec :: GeomAlgebra g mv => Int -> [Double] -> mv

Returns a normalized k-vector with given grade and coefficients.

Parts

getGrade :: CliffAlgebra g mv => Int -> mv -> mv

Extracts k-blade from a multivector.

trace :: CliffAlgebra g mv => mv -> Double

Extracts a scalar part from a multivector.

weight :: CliffAlgebra g mv => mv -> mv

Extracts vanishing part from a multivector.

bulk :: CliffAlgebra g mv => mv -> mv

Extracts non-vanishing part from a multivector.

components :: CliffAlgebra g mv => mv -> [mv]

Returns a list of monomial components in the multivector.

Predicates

isScalar :: CliffAlgebra g mv => mv -> Bool

Returns True if multivector is scalar and False otherwise.

isHomogeneous :: CliffAlgebra g mv => mv -> Bool

Returns True if multivector is a k-vector and False otherwise..

isSingular :: (Num mv, CliffAlgebra g mv) => mv -> Bool

Returns True if multivector is singular (non-invertible) and False otherwise.

Normalization

norm :: CliffAlgebra g mv => mv -> Double

Returns norm of a multivector.

norm2 :: CliffAlgebra g mv => mv -> Double

Returns square norm of a multivector.

normalize :: CliffAlgebra g mv => mv -> mv

Returns normalized multivector for given nonsingular multivector.

Inversion

isInvertible :: CliffAlgebra g mv => mv -> Bool

Returns True if multivector is not singular (invertible) and False otherwise.

reciprocal :: CliffAlgebra g mv => mv -> mv

Returns reciprocal for nonsingular k-vector.

Geometric algebras

class CliffAlgebra g mv => GeomAlgebra g mv where

The class representing Clifford algebra that have geometric representation.

Methods

point :: [Double] -> mv

Representation of a point, given as coordinate list, as a k-vector.

coord :: mv -> [Double]

Representation of a k-vector as a point, given as coordinate list.

spaceDim :: mv -> Int

Returns a geometric dimention for the embedding space of the algebra. The multivector is used as a type witness.

dim :: mv -> Int

Returns a geometric dimention for a multivector. It does not alwais correspond to grade, and depends on the dimension of the embedding space for dual algebras.

Elements

pseudoScalar :: CliffAlgebra g mv => mv

Returns a pseudoscalar of the algebra.

>>> pseudoScalar :: VGA 3
e₁₂₃
>>> pseudoScalar :: PGA 3
e₀₁₂₃

basis :: CliffAlgebra g mv => [mv]

Returns a list of all monomial components in the algebra.

>>> basis :: [VGA 3]
[1.0,e₁,e₂,e₃,e₁₂,e₁₃,e₂₃,e₁₂₃]
>>> basis :: [PGA 3]
[1.0,e₀,e₁,e₂,e₃,e₀₁,e₀₂,e₀₃,e₁₂,e₁₃,e₂₃,e₀₁₂,e₀₁₃,e₀₂₃,e₁₂₃,e₀₁₂₃]

Predicates

isPoint :: GeomAlgebra g mv => mv -> Bool

Returns True if k-vector represents a point.

isLine :: GeomAlgebra g mv => mv -> Bool

Returns True if k-vector represents a line.

isPlane :: GeomAlgebra g mv => mv -> Bool

Returns True if k-vector represents a plane.

Geometric tools

(∨) :: GeomAlgebra g mv => mv -> mv -> mv infixr 9

Regressive product of two multivectors.

meet :: GeomAlgebra g mv => mv -> mv -> mv

Normalized outer product of two multivectors.

join :: GeomAlgebra g mv => mv -> mv -> mv

Normalized regressive product of two multivectors.

projectOn :: GeomAlgebra g mv => mv -> mv -> mv

Returns projection of object a on object b.

(->|) :: GeomAlgebra g mv => mv -> mv -> mv

Infix operator for projection.

antiprojectTo :: GeomAlgebra g mv => mv -> mv -> mv

Returns antiprojection of object a on object b.

(<-|) :: GeomAlgebra g mv => mv -> mv -> mv

Infix operator for antiprojection.

segmentMeet :: GeomAlgebra g mv => mv -> (mv, mv) -> Maybe mv

Returns intersection of an object and a segment.

reflectAt :: (Num mv, GeomAlgebra g mv) => mv -> mv -> mv

Returns reflection of object a against object b.

rotateAt :: (Num mv, GeomAlgebra g mv) => mv -> Double -> mv -> mv

Rotates object x against the object p by given angle.

angle :: GeomAlgebra g mv => mv -> mv -> Double

Returns an angle between two multivectors.

shiftAlong :: (GeomAlgebra g mv, Fractional mv) => mv -> mv -> mv

Translates an object along the object l by distance, given by norm of l.

shiftAlong' :: (GeomAlgebra g mv, Fractional mv) => mv -> Double -> mv -> mv

Translates object x along the object l by given distance d.

rescale :: (Num mv, CliffAlgebra g mv) => Double -> mv -> mv

Rescales an object a by given value.

stretch :: (Num mv, CliffAlgebra g mv) => Double -> mv -> mv

Rescales magnitude of object a by given value.

Types

data Dual a

Wrapper which represents dual algebra for a given one.

type family Cl (p :: Nat) (q :: Nat) (r :: Nat) where ...

Type family for Clifford algebras with given signature.

Equations

Cl 0 0 0 = Double

Cl 0 n 0 = Outer n

Cl n 0 0 = VGA n

Cl 2 1 0 = PGA 2

Cl 3 1 0 = PGA 3

Cl n 1 0 = PGA n

Cl p q r = CA p q r

newtype Outer (n :: Nat)

Outer (Grassmann) algebra of given dimension.

Constructors

Outer MapLS

type Outer' n = Dual (Outer n)

Dual outer (Grassmann) algebra of given dimension.

newtype VGA (n :: Nat)

Affine vector geometric algebra of given dimension.

Constructors

VGA (CA n 0 0)

type VGA' n = Dual (VGA n)

Dual affine vector geometric algebra of given dimension.

newtype PGA (n :: Nat)

Projective geometric algebra of given dimension.

Constructors

PGA (CA n 1 0)

type PGA' n = Dual (PGA n)

Dual projective geometric algebra of given dimension.

Effective realisations of geometric algebras

newtype VGA2

Tabulated 2D affine geometric algebra.

Constructors

VGA2 MapLS

newtype VGA3

Tabulated 3D affine geometric algebra.

Constructors

VGA3 MapLS

newtype PGA2

Tabulated 2D projective geometric algebra.

Constructors

PGA2 MapLS

newtype PGA3

Tabulated 3D projective geometric algebra.

Constructors

PGA3 MapLS

newtype PGA4

Tabulated 4D projective geometric algebra.

Constructors

PGA4 MapLS

defineElements :: CliffAlgebra Int a => [a] -> Q [Dec]

Template which generates aliases for given basis of Clifford algebra.

tabulateGA :: String -> Integer -> String -> Q Decs

Template which generates instanses for tabulated geometric algebra.

Basis elements for geometric algebras.

e0 :: CliffAlgebra Int a => a

e1 :: CliffAlgebra Int a => a

e2 :: CliffAlgebra Int a => a

e3 :: CliffAlgebra Int a => a

e4 :: CliffAlgebra Int a => a

e01 :: CliffAlgebra Int a => a

e02 :: CliffAlgebra Int a => a

e03 :: CliffAlgebra Int a => a

e04 :: CliffAlgebra Int a => a

e12 :: CliffAlgebra Int a => a

e13 :: CliffAlgebra Int a => a

e14 :: CliffAlgebra Int a => a

e23 :: CliffAlgebra Int a => a

e24 :: CliffAlgebra Int a => a

e34 :: CliffAlgebra Int a => a

e012 :: CliffAlgebra Int a => a

e013 :: CliffAlgebra Int a => a

e014 :: CliffAlgebra Int a => a

e023 :: CliffAlgebra Int a => a

e024 :: CliffAlgebra Int a => a

e034 :: CliffAlgebra Int a => a

e123 :: CliffAlgebra Int a => a

e124 :: CliffAlgebra Int a => a

e134 :: CliffAlgebra Int a => a

e234 :: CliffAlgebra Int a => a

e1234 :: CliffAlgebra Int a => a

e0234 :: CliffAlgebra Int a => a

e0134 :: CliffAlgebra Int a => a

e0124 :: CliffAlgebra Int a => a

e0123 :: CliffAlgebra Int a => a

e01234 :: CliffAlgebra Int a => a