Algeometry.GeometricAlgebra

Description

Documentation

class Eq b => LinSpace b el | el -> b where

The class representing a general linear space with basis of type b and element (vector) of type el.

Minimal complete definition

zero, isZero, monom, add, lmap, lapp, lfilter, assoc

Methods

zero :: el

The zero element.

isZero :: el -> Bool

Predicate for the zero element.

monom :: b -> Double -> el

Constructor for the monom.

isMonom :: el -> Bool

Predicate for the monom.

add :: el -> el -> el

Addition for two elements.

lmap :: (b -> Maybe (b, Double)) -> el -> el

Mapping of partial unary linear function through an element.

lapp :: (b -> b -> Maybe (b, Double)) -> el -> el -> el

Distribution of partial binary linear function through two elements.

lfilter :: (b -> Double -> Bool) -> el -> el

Extraction of elements with given predicate.

coeff :: b -> el -> Double

Scalar coefficient of an element

assoc :: el -> [(b, Double)]

Representation of an element as assoclist.

class (Eq mv, Ord g, LinSpace [g] mv) => CliffAlgebra g mv where

The class representing Clifford algebra with generators of type g and element (multivector) of type mv.

Minimal complete definition

algebraSignature, square, generators

Methods

algebraSignature :: mv -> (Int, Int, Int)

The signature of Clifford algebra. The first argument is proxy needed to resolve the functional dependence for class instance.

square :: mv -> g -> Double

Squares of the generators, which define the Clifford algebra. The first argument is proxy needed to resolve the functional dependence for class instance.

generators :: [mv]

List of generators of the Clifford algebra.

grade :: mv -> Int

Returns the grade of a multivector.

geom :: mv -> mv -> mv

The geometric product.

outer :: mv -> mv -> mv

The outer product (same as (∧)).

rcontract :: mv -> mv -> mv

The right contraction (same as (|-)).

lcontract :: mv -> mv -> mv

The left contraction (same as (-|)).

inner :: mv -> mv -> mv

The inner product (same as (·)).

rev :: mv -> mv

The reverse of a multivector.

inv :: mv -> mv

The inversion of a multivector.

conj :: mv -> mv

The conjugate of a multivector

dual :: mv -> mv

The dual of a multivector.

rcompl :: mv -> mv

The right complement of a multivector.

lcompl :: mv -> mv

The left complement of a multivector.

class CliffAlgebra g mv => GeomAlgebra g mv where

The class representing Clifford algebra that have geometric representation.

Methods

point :: [Double] -> mv

Representation of a point, given as coordinate list, as a k-vector.

coord :: mv -> [Double]

Representation of a k-vector as a point, given as coordinate list.

spaceDim :: mv -> Int

Returns a geometric dimention for the embedding space of the algebra. The multivector is used as a type witness.

dim :: mv -> Int

Returns a geometric dimention for a multivector. It does not alwais correspond to grade, and depends on the dimension of the embedding space for dual algebras.

e :: CliffAlgebra g mv => g -> mv

Returns a generator of the algebra.

e_ :: CliffAlgebra g mv => [g] -> mv

Returns a monomial element of the algebra (outer product of generators).

scalar :: CliffAlgebra g mv => Double -> mv

Returns a scalar element of the algebra.

>>> scalar (sin 1) + scalar (cos 1) * e12 :: VGA 2
0.8414709848078965 + 0.5403023058681398e₁₂

kvec :: GeomAlgebra g mv => Int -> [Double] -> mv

Returns a k-vector with given grade and coefficients.

>>> kvec 1 [1,2,3] :: VGA 2
e₁ + 2.0e₂ + 3.0e₃
>>> kvec 2 [1,2,3] :: VGA 3
e₁₂ + 2.0e₁₃ + 3.0e₂₃

nvec :: GeomAlgebra g mv => Int -> [Double] -> mv

Returns a normalized k-vector with given grade and coefficients.

avec :: GeomAlgebra g mv => Int -> [Double] -> mv

Returns a normalized k-antivector with given grade and coefficients.

>>> avec 1 [1,2,3] :: VGA 3
3.0e₁₂ + 2.0e₁₃ + e₂₃
>>> avec 2 [1,2,3] :: VGA 3
3.0e₁ + 2.0e₂ + e₃

angle :: GeomAlgebra g mv => mv -> mv -> Double

Returns an angle between two multivectors.

elems :: LinSpace b el => el -> [b]

Returns list of basis elements for an element.

coefs :: LinSpace b el => el -> [Double]

Returns list of coefficients for an element.

terms :: LinSpace b el => el -> [el]

Returns list of monoms for an element.

lerp :: (Num el, LinSpace b el) => el -> el -> Double -> el

Linear interpolation for two elements, parameterized for interval [0,1].

isScalar :: CliffAlgebra g mv => mv -> Bool

Returns True if multivector is scalar and False otherwise.

isHomogeneous :: CliffAlgebra g mv => mv -> Bool

Returns True if multivector is a k-vector and False otherwise..

isSingular :: (Num mv, CliffAlgebra g mv) => mv -> Bool

Returns True if multivector is singular (non-invertible) and False otherwise.

trace :: CliffAlgebra g mv => mv -> Double

Extracts a scalar part from a multivector.

getGrade :: CliffAlgebra g mv => Int -> mv -> mv

Extracts k-blade from a multivector.

components :: CliffAlgebra g mv => mv -> [mv]

Returns a list of monomial components in the multivector.

pseudoScalar :: CliffAlgebra g mv => mv

Returns a pseudoscalar of the algebra.

>>> pseudoScalar :: VGA 3
e₁₂₃
>>> pseudoScalar :: PGA 3
e₀₁₂₃

basis :: CliffAlgebra g mv => [mv]

Returns a list of all monomial components in the algebra.

>>> basis :: [VGA 3]
[1.0,e₁,e₂,e₃,e₁₂,e₁₃,e₂₃,e₁₂₃]
>>> basis :: [PGA 3]
[1.0,e₀,e₁,e₂,e₃,e₀₁,e₀₂,e₀₃,e₁₂,e₁₃,e₂₃,e₀₁₂,e₀₁₃,e₀₂₃,e₁₂₃,e₀₁₂₃]

isInvertible :: CliffAlgebra g mv => mv -> Bool

Returns True if multivector is not singular (invertible) and False otherwise.

reciprocal :: CliffAlgebra g mv => mv -> mv

Returns reciprocal for nonsingular k-vector.

scale :: LinSpace b el => Double -> el -> el

Returns an element of a linear space, scaled by given factor.

weight :: CliffAlgebra g mv => mv -> mv

Extracts vanishing part from a multivector.

bulk :: CliffAlgebra g mv => mv -> mv

Extracts non-vanishing part from a multivector.

norm :: CliffAlgebra g mv => mv -> Double

Returns norm of a multivector.

norm2 :: CliffAlgebra g mv => mv -> Double

Returns square norm of a multivector.

normalize :: CliffAlgebra g mv => mv -> mv

Returns normalized multivector for given nonsingular multivector.

(∧) :: CliffAlgebra g mv => mv -> mv -> mv infixr 9

The infix operator for the outer product

(∨) :: GeomAlgebra g mv => mv -> mv -> mv infixr 9

Regressive product of two multivectors.

(|-) :: CliffAlgebra g mv => mv -> mv -> mv infix 8

The infix operator for rcontract.

(-|) :: CliffAlgebra g mv => mv -> mv -> mv infix 8

The infix operator for the lcontract.

(∙) :: CliffAlgebra g mv => mv -> mv -> mv infix 8

The infix operator for the inner product

(•) :: CliffAlgebra g mv => mv -> mv -> Double infix 9

The infix operator for the scalar product

(->|) :: GeomAlgebra g mv => mv -> mv -> mv

Infix operator for projection.

(<-|) :: GeomAlgebra g mv => mv -> mv -> mv

Infix operator for antiprojection.

meet :: GeomAlgebra g mv => mv -> mv -> mv

Normalized outer product of two multivectors.

join :: GeomAlgebra g mv => mv -> mv -> mv

Normalized regressive product of two multivectors.

segmentMeet :: GeomAlgebra g mv => mv -> (mv, mv) -> Maybe mv

Returns intersection of an object and a segment.

clipPoly :: GeomAlgebra g mv => [mv] -> mv -> [mv]

For polygon and a line returns list of intersection points.

reflectAt :: (Num mv, GeomAlgebra g mv) => mv -> mv -> mv

Returns reflection of object a against object b.

rotateAt :: (Num mv, GeomAlgebra g mv) => mv -> Double -> mv -> mv

Rotates object x against the object p by given angle.

projectOn :: GeomAlgebra g mv => mv -> mv -> mv

Returns projection of object a on object b.

antiprojectTo :: GeomAlgebra g mv => mv -> mv -> mv

Returns antiprojection of object a on object b.

shiftAlong :: (GeomAlgebra g mv, Fractional mv) => mv -> mv -> mv

Translates an object along the object l by distance, given by norm of l.

shiftAlong' :: (GeomAlgebra g mv, Fractional mv) => mv -> Double -> mv -> mv

Translates object x along the object l by given distance d.

rescale :: (Num mv, CliffAlgebra g mv) => Double -> mv -> mv

Rescales an object a by given value.

stretch :: (Num mv, CliffAlgebra g mv) => Double -> mv -> mv

Rescales magnitude of object a by given value.

isPoint :: GeomAlgebra g mv => mv -> Bool

Returns True if k-vector represents a point.

isLine :: GeomAlgebra g mv => mv -> Bool

Returns True if k-vector represents a line.

isPlane :: GeomAlgebra g mv => mv -> Bool

Returns True if k-vector represents a plane.