Algeometry.GeometricNum

Description

Documentation

newtype GeometricNum a

This newtype implements various classes for arbitrary Clifford algebra. It may be used to derive numeric classes for exact Geometric algebra implementation using deriving via mechanics.

Constructors

GeometricNum a