Algeometry.SVG

Description

Types

data Fig

Represents graphic primitives

Constructors

Point [Attribute] XY

Line [Attribute] XY XY

Polygon [Attribute] [XY]

data Figure a b

Representation of a figure with many graphic objects with attributes.

figure :: Figure PGA2 b -> [Fig]

Returns a figure as a list of primitives.

getFigure :: Figure a b -> [([a], [Attribute])]

Returns list of graphic objects with attributes.

getResult :: Figure a b -> b

Returns value, stored in Figure type.

mapFig :: (a1 -> a2) -> Figure a1 b -> Figure a2 b

Applies transformation to objects, stored in Figure type.

data Animation a

Type representing animationa frames.

SVG generation and output

put :: a -> Figure a a

Adds multivector as graphic object to Figure.

(@) :: a -> [Attribute] -> Figure a a infix 1

Adds multivector as graphic object with given attributes.

(<@) :: Figure a b -> [Attribute] -> Figure a b infix 1

Appends attributes to attributed object.

svg :: Monad m => [Fig] -> HtmlT m ()

Returns raw SVG for a figure.

writeSVG :: FilePath -> Figure PGA2 b -> IO ()

Outputs figure to SVG file.

writePNG :: String -> Figure PGA2 b -> IO ()

Outputs figure to PNG file.

Simple animation

writeFrame :: Int -> FilePath -> Animation a -> IO ()

Outputs single frame of animation, given by a frame number.

runAnimation :: String -> Animation a -> IO ()

Outputs animation as GIF file, using convert utility.

animate :: Int -> (Double, Double) -> (Double -> Figure PGA2 a) -> Animation ()

Creates animation from a parameterized function and given nuber of frames.

transform :: Int -> (Double -> Figure PGA2 a) -> Animation ()

Creates animation with nonlinear paremeterization.

animateList :: [Figure PGA2 a] -> Animation ()

Creates animation from a list of figures.

Constructors

origin :: GeomAlgebra b a => a

Returns multivector, which represents the origin.

point :: GeomAlgebra b a => [Double] -> a

Returns multivector, which represents a point.

line :: GeomAlgebra b a => [Double] -> [Double] -> a

Returns multivector, which represents a line.

nline :: GeomAlgebra b a => [Double] -> [Double] -> a

Returns multivector, which represents a normalized line.

vect :: GeomAlgebra b a => [Double] -> a

Returns multivector, which represents a vector (line passing through the origin).

nvect :: GeomAlgebra b a => [Double] -> a

Returns multivector, which represents a normalized vector (line passing through the origin).

segm :: GeomAlgebra b a => [Double] -> [Double] -> Figure a [a]

Returns a graphic object, which represents a segment given by coordinates.

polygon :: GeomAlgebra e a => [a] -> Figure a [a]

Returns a graphic object, which represents a polygon.

polyline :: GeomAlgebra e a => [a] -> Figure a [a]

Returns a graphic object, which represents a polyline.

regularPoly :: GeomAlgebra e a => Double -> Figure a [a]

Returns a graphic object, which represents a regular polygon.

segment :: GeomAlgebra e a => a -> a -> Figure a [a]

Returns a graphic object, which represents a segment? given by two points.

plane :: GeomAlgebra b a => [Double] -> [Double] -> [Double] -> a

Returns multivector, which represents a plane passing through three points.

plane2 :: (Fractional a, GeomAlgebra e a) => a -> a -> (Double, Double) -> Figure a a

Returns graphic object, which represents a plane passing through a point and a line.

plane3 :: (Fractional a, GeomAlgebra e a) => a -> a -> a -> (Double, Double) -> Figure a a

Returns graphic object, which represents a plane passing through three points.

orthoPlane :: (Fractional a, GeomAlgebra e a) => a -> a -> (Double, Double) -> Figure a a

Returns graphic object, which represents a plane passing through a point, orthogonal to a line.

label :: GeomAlgebra b a => [Double] -> String -> Figure a a

Returns a label at given position.

display :: (Show s, GeomAlgebra b a) => [Double] -> s -> Figure a a

Returns a label with given value.

Miscellaneous tools

viewPoint :: [Double] -> Figure PGA3 b -> Figure PGA2 b

Projects 3D points to a 2D plane, using given viewpoint.

axis :: Figure PGA2 PGA2

Adds 2D axes to a Figure

grid :: Figure PGA2 ()

Adds 2D unit grid to a Figure

background :: GeomAlgebra e a => String -> Figure a [a]

Adds 2D background rectangle to a Figure

clipPoly :: [PGA2] -> PGA2 -> [PGA2]

For polygon and a line returns list of intersection points.

Orphan instances

With [Attribute]

Instance details

Methods

with :: [Attribute] -> [Attribute] -> [Attribute]