Algeometry.Types

Description

Documentation

newtype CA (p :: Nat) (q :: Nat) (r :: Nat)

Constructors

CA MapLS

newtype Dual a

Wrapper which represents dual algebra for a given one.

Constructors

Dual

Fields

getDual :: a

newtype Outer (n :: Nat)

Outer (Grassmann) algebra of given dimension.

Constructors

Outer MapLS

type Outer' n = Dual (Outer n)

Dual outer (Grassmann) algebra of given dimension.

newtype VGA (n :: Nat)

Affine vector geometric algebra of given dimension.

Constructors

VGA (CA n 0 0)

type VGA' n = Dual (VGA n)

Dual affine vector geometric algebra of given dimension.

newtype PGA (n :: Nat)

Projective geometric algebra of given dimension.

Constructors

PGA (CA n 1 0)

type PGA' n = Dual (PGA n)

Dual projective geometric algebra of given dimension.

class CliffAlgebra a => Tabulated a where

Methods

squareT :: a -> Generator a -> Double

signatureT :: a -> (Int, Int, Int)

generatorsT :: [a]

indexT :: IndexMap a

geomT :: Table a (Int, Int)

outerT :: Table a (Int, Int)

innerT :: Table a (Int, Int)

lcontractT :: Table a (Int, Int)

rcontractT :: Table a (Int, Int)

revT :: Table a Int

invT :: Table a Int

conjT :: Table a Int

dualT :: Table a Int

rcomplT :: Table a Int

lcomplT :: Table a Int

newtype TabulatedGA a

Constructors

TabulatedGA a

type MapLS = Map [Int] Double

Representation of linear space as a map, indexed by integer indexes.

defineElements :: (Generator a ~ Int, CliffAlgebra a) => [a] -> Q [Dec]

Template which generates aliases for given basis of Clifford algebra.

tabulateGA :: String -> Integer -> String -> Q Decs

Template which generates instanses for tabulated geometric algebra.

Orphan instances

Ord e => LinSpace (Map e Double)

Instance details

Methods

zero :: Map e Double

isZero :: Map e Double -> Bool

monom :: Basis (Map e Double) -> Double -> Map e Double

isMonom :: Map e Double -> Bool

add :: Map e Double -> Map e Double -> Map e Double

lmap :: (Basis (Map e Double) -> Maybe (Basis (Map e Double), Double)) -> Map e Double -> Map e Double

lapp :: (Basis (Map e Double) -> Basis (Map e Double) -> Maybe (Basis (Map e Double), Double)) -> Map e Double -> Map e Double -> Map e Double

lfilter :: (Basis (Map e Double) -> Double -> Bool) -> Map e Double -> Map e Double

coeff :: Basis (Map e Double) -> Map e Double -> Double

assoc :: Map e Double -> [(Basis (Map e Double), Double)]