Algeometry.Types

Description

Documentation

newtype Pos a (p :: Nat)

Constructors

Pos a

newtype Neg a (r :: Nat)

Constructors

Neg a

newtype Zero a (q :: Nat)

Constructors

Zero a

newtype CA (p :: Nat) (q :: Nat) (r :: Nat)

Constructors

CA (Pos (Zero (Neg MapLS r) q) p)

newtype Outer (n :: Nat)

Outer (Grassmann) algebra of given dimention.

Constructors

Outer MapLS

newtype VGA (n :: Nat)

Affine vector geometric algebra of given dimention.

Constructors

VGA (CA n 0 0)

newtype PGA (n :: Nat)

Projective geometric algebra of given dimention.

Constructors

PGA (CA n 1 0)

data GeometricNum a

class Tabulated e a | a -> e where

Methods

squareT :: a -> e -> Double

signatureT :: a -> (Int, Int, Int)

generatorsT :: [a]

indexT :: IndexMap [e] a

geomT :: Table e a (Int, Int)

outerT :: Table e a (Int, Int)

innerT :: Table e a (Int, Int)

lcontractT :: Table e a (Int, Int)

rcontractT :: Table e a (Int, Int)

revT :: Table e a Int

invT :: Table e a Int

conjT :: Table e a Int

dualT :: Table e a Int

rcomplT :: Table e a Int

lcomplT :: Table e a Int

newtype TabulatedGA a

Constructors

TabulatedGA a

type MapLS = Map [Int] Double

Representation of linear space as a map, indexed by integer indexes.

defineElements :: CliffAlgebra Int a => [a] -> Q [Dec]

Template which generates aliases for given basis of Clifford algebra.

tabulateGA :: String -> Integer -> Q Decs

Template which generates instanses for tabulated geometric algebra.

Orphan instances

Ord e => LinSpace e (Map e Double)

Instance details

Methods

zero :: Map e Double

isZero :: Map e Double -> Bool

monom :: e -> Double -> Map e Double

isMonom :: Map e Double -> Bool

add :: Map e Double -> Map e Double -> Map e Double

lmap :: (e -> Maybe (e, Double)) -> Map e Double -> Map e Double

lapp :: (e -> e -> Maybe (e, Double)) -> Map e Double -> Map e Double -> Map e Double

lfilter :: (e -> Double -> Bool) -> Map e Double -> Map e Double

coeff :: e -> Map e Double -> Double

assoc :: Map e Double -> [(e, Double)]